requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
asyncio>=3.4.3
pydantic>=1.8.2
email-validator>=1.1.3
typing-extensions>=4.0.0
starlette>=0.14.2
requests>=2.26.0